feature: publish week

+++ backend/src/usecases/shiftUsecase.ts
@@ -1,9 +1,9 @@
 import * as shiftRepository from "../database/default/repository/shiftRepository";
 import { FindManyOptions, FindOneOptions } from "typeorm";
 import Shift from "../database/default/entity/shift";
-import { ICreateShift, IUpdateShift, WeekStatus} from "../shared/interfaces";
+import { ICreateShift, IUpdateShift, WeekStatus } from "../shared/interfaces";
 import { HttpError } from "../shared/classes/HttpError";
-import * as weekUsecase from "./weekUsecase"
+import * as weekUsecase from "./weekUsecase";
 import moment from "moment";
 import Week from "../database/default/entity/week";
 
@@ -25,12 +25,12 @@ export const create = async (payload: ICreateShift): Promise<Shift> => {
     throw new HttpError(400, "Shift is overlapping");
   }
 
-  const weekNo = moment(payload.date, 'DD-MM-YYYY').isoWeek()
+  const weekNo = moment(payload.date, "DD-MM-YYYY").isoWeek();
   let week: Week = await weekUsecase.findOne({
-    week: weekNo
-  })
+    week: weekNo,
+  });
 
-  if(!week) {
+  if (!week) {
     let [startDate, endDate] = weekUsecase.getWeekRange(payload.date);
     week = new Week();
     week.startDate = startDate;
@@ -40,8 +40,8 @@ export const create = async (payload: ICreateShift): Promise<Shift> => {
     week = await weekUsecase.create(week);
   }
 
-  if (week.status === WeekStatus.PUBLISHED){
-    throw new HttpError(400, 'Cannot create shift in published week')
+  if (week.status === WeekStatus.PUBLISHED) {
+    throw new HttpError(400, "Cannot create shift in published week");
   }
 
   const shift = new Shift();
@@ -50,7 +50,7 @@ export const create = async (payload: ICreateShift): Promise<Shift> => {
   shift.startTime = payload.startTime;
   shift.endTime = payload.endTime;
   shift.week = week;
-
+  
   return shiftRepository.create(shift);
 };
 
@@ -60,13 +60,13 @@ export const updateById = async (
 ): Promise<Shift> => {
   let shift: Shift = await findById(id);
 
-  if(!shift) {
-    throw new HttpError(404, "Shift not found")
+  if (!shift) {
+    throw new HttpError(404, "Shift not found");
   }
 
-  // if (shift.status === ShiftStatus.PUBLISHED) {
-  //   throw new HttpError(400, "Published shift cannot be updated");
-  // }
+  if (shift.week.status === WeekStatus.PUBLISHED) {
+    throw new HttpError(400, "Published shift cannot be updated");
+  }
 
   let isOverlapping = await checkOverlappingShift(shift);
 
@@ -83,27 +83,29 @@ export const deleteById = async (id: string | string[]) => {
   if (typeof id === "string") {
     let shift: Shift = await findById(id);
 
-    if(!shift) {
-      throw new HttpError(400, "Shift not found")
+    if (!shift) {
+      throw new HttpError(400, "Shift not found");
     }
 
-    // if (shift.status === ShiftStatus.PUBLISHED) {
-    //   throw new HttpError(400, "Published shift cannot be deleted")
-    // }
+    if (shift.week.status === WeekStatus.PUBLISHED) {
+      throw new HttpError(400, "Published shift cannot be updated");
+    }
     return shiftRepository.deleteById(id);
   } else {
     id.map(async (entry) => {
       let shift: Shift = await findById(entry);
-      // if (shift.status === ShiftStatus.PUBLISHED) {
-      //   throw new HttpError(400, "Published shift cannot be deleted")
-      // }
+      if (shift.week.status === WeekStatus.PUBLISHED) {
+        throw new HttpError(400, "Published shift cannot be updated");
+      }
     });
   }
 
   return shiftRepository.deleteById(id);
 };
 
-export const checkOverlappingShift = async (shift: Shift | ICreateShift): Promise<Boolean> => {
+export const checkOverlappingShift = async (
+  shift: Shift | ICreateShift
+): Promise<Boolean> => {
   const todayShifts = await find({
     where: {
       date: shift.date,
@@ -111,13 +113,17 @@ export const checkOverlappingShift = async (shift: Shift | ICreateShift): Promis
   });
 
   for (let i = 0; i < todayShifts.length; i++) {
-    let startTimeOverlap = shift.startTime > todayShifts[i].startTime && shift.startTime < todayShifts[i].endTime
-    let endTimeOverlap = shift.endTime > todayShifts[i].startTime && shift.endTime < todayShifts[i].endTime
-    
+    let startTimeOverlap =
+      shift.startTime > todayShifts[i].startTime &&
+      shift.startTime < todayShifts[i].endTime;
+    let endTimeOverlap =
+      shift.endTime > todayShifts[i].startTime &&
+      shift.endTime < todayShifts[i].endTime;
+
     if (startTimeOverlap || endTimeOverlap) {
       return true;
     }
   }
 
   return false;
-}
+};
